backend/pyproject.toml
[project]
name = "autobook-backend"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.9",
    "greenlet>=3.0.0",
    "email-validator>=2.0.0",
]
